Create appropriate sidecars for store generated values (Issue #1303)
The code creates a sidecar for properties that may get a store generated value because a temporary value generator is used either directly or for the root principal in the case of FKs, or the property or the property is directly marked to be computed or get a store default. Note that the actual value of the property is not used (as it was previously) since it can change and hence a store value may look like it is not needed when creating the sidecar, but then it may actually be needed if the value later changes later.

This required that value generators be changed to indicate whether or not they generate temporary values without actually generating a value. This means a value generator must either generate temps or not and cannot switch, but this seems like a reasonable limitation.